[Feat] 홈 - 오늘의 일정 아이템에 담당자 표시 (#309)
* [Feat] 홈 - 오늘의 일정 아이템 담당자 컴포넌트 추가 (#296)

* chore : string 리소스 추가

* feat : 홈 - 오늘의 일정 아이템에 담당자 컴포넌트 구현

* chore : 코드 포맷팅 적용 및 홈 컴포넌트 패키지 분류

* [Feat, Refactor] 홈 - State 코드 최적화 및 오늘의 일정 아이템에 일정 담당자 추가 (#305)

* refactor : 컬러 토큰값으로 변경

* chore, refactor : TodoItem Immutable 어노테이션 제거 및 TodoState -> TodoItem으로 UI 모델명 변경

* refactor : BalanceGameState 코드 최적화

* feat : 일정 아이템에 일정 담당자 추가

* [Feat] ContentRole 서버 DTO 추가 (#307)

* chore : 사용하지않는 서버 DTO 제거

* feat : ContentAsignee DTO 추가

* chore : fixme 주석 추가

* feat : ContentRole 필드 서버와 동일하게 변경

* chore : 코드 포맷팅 적용

* refactor : ContentAsignee DTO 명칭 변경

* chore : 도메인 DTO ContentRole 이름 변경

* chore : contentAsignee 철자 오류 수정

* chore : ContentRole -> ContentAsignee 로 네이밍 변경
