+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/repository/CalendarRepositoryImpl.kt
@@ -72,7 +72,12 @@ class CalendarRepositoryImpl(
         userTimezone: String?,
     ): List<Todo> =
         safeCall {
-            remoteCalendarDataSource.getSchedules(startDate, endDate, userTimezone).toTodo()
+            remoteCalendarDataSource
+                .getSchedules(
+                    startDate = startDate,
+                    endDate = endDate,
+                    userTimeZone = userTimezone,
+                ).toTodo()
         }
 
     override suspend fun getHolidays(year: Int): List<Holiday> =
