+++ feature/home/src/androidMain/kotlin/com/whatever/caramel/feature/home/HomeScreenPreviewData.kt
@@ -1,8 +1,11 @@
 package com.whatever.caramel.feature.home
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import com.whatever.caramel.core.domain.vo.content.ContentAssignee
 import com.whatever.caramel.feature.home.mvi.HomeState
-import com.whatever.caramel.feature.home.mvi.TodoState
+import com.whatever.caramel.feature.home.mvi.TodoItem
+import kotlinx.collections.immutable.persistentListOf
+import kotlinx.collections.immutable.toImmutableList
 
 internal class HomeScreenPreviewData : PreviewParameterProvider<HomeState> {
     override val values: Sequence<HomeState> =
@@ -10,14 +13,14 @@ internal class HomeScreenPreviewData : PreviewParameterProvider<HomeState> {
             HomeState(
                 daysTogether = 0,
                 shareMessage = "",
-                todos = emptyList(),
+                todoList = persistentListOf(),
                 isShowBottomSheet = true,
                 isLoading = false,
             ),
             HomeState(
                 daysTogether = 0,
                 shareMessage = "",
-                todos = emptyList(),
+                todoList = persistentListOf(),
                 isShowBottomSheet = false,
                 isLoading = false,
                 coupleState = HomeState.CoupleState.DISCONNECT,
@@ -25,35 +28,58 @@ internal class HomeScreenPreviewData : PreviewParameterProvider<HomeState> {
             HomeState(
                 daysTogether = 0,
                 shareMessage = "",
-                todos = emptyList(),
+                todoList = persistentListOf(),
                 isShowBottomSheet = false,
                 isLoading = false,
             ),
             HomeState(
                 daysTogether = 1120,
                 shareMessage = "내일 여기서 만나",
-                todos = emptyList(),
+                todoList = persistentListOf(),
                 isShowBottomSheet = false,
                 isLoading = false,
             ),
             HomeState(
                 daysTogether = 1120,
                 shareMessage = "내일 여기서 만나",
-                todos = listOf(TodoState(id = 1, title = "오늘 캘린더에 첫 번째로 적혀있는 것이 표시됩니다. 글자수가 넘어가면 어떻게 되나요?")),
+                todoList =
+                    listOf(
+                        TodoItem(
+                            id = 1,
+                            title = "오늘 캘린더에 첫 번째로 적혀있는 것이 표시됩니다. 글자수가 넘어가면 어떻게 되나요?",
+                            role = ContentAssignee.ME,
+                        ),
+                    ).toImmutableList(),
                 isShowBottomSheet = false,
                 isLoading = false,
             ),
             HomeState(
                 daysTogether = 1120,
                 shareMessage = "내일 여기서 만나",
-                todos = (0..7).map { TodoState(it.toLong(), "오늘 캘린더에 첫 번째로 적혀있는 것이 표시됩니다. 글자수가 넘어가면 어떻게 되나요?") },
+                todoList =
+                    (0..7)
+                        .map {
+                            TodoItem(
+                                id = it.toLong(),
+                                title = "오늘 캘린더에 첫 번째로 적혀있는 것이 표시됩니다. 글자수가 넘어가면 어떻게 되나요?",
+                                role = ContentAssignee.US,
+                            )
+                        }.toImmutableList(),
                 isShowBottomSheet = false,
                 isLoading = false,
             ),
             HomeState(
                 daysTogether = 1120,
                 shareMessage = "내일 여기서 만나",
-                todos = (0..7).map { TodoState(it.toLong(), "오늘 캘린더에 첫 번째로 적혀있는 것이 표시됩니다. 글자수가 넘어가면 어떻게 되나요?") },
+                todoList =
+                    (0..7)
+                        .map {
+                            TodoItem(
+                                id = it.toLong(),
+                                title = "오늘 캘린더에 첫 번째로 적혀있는 것이 표시됩니다. 글자수가 넘어가면 어떻게 되나요?",
+                                role = ContentAssignee.PARTNER,
+                            )
+                        }.toImmutableList(),
                 isShowBottomSheet = false,
                 isLoading = true,
             ),
