+++ feature/home/src/androidMain/kotlin/com/whatever/caramel/feature/home/QuizPreviewData.kt
@@ -2,77 +2,79 @@ package com.whatever.caramel.feature.home
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import com.whatever.caramel.core.domain.vo.user.Gender
-import com.whatever.caramel.feature.home.mvi.BalanceGameOptionState
-import com.whatever.caramel.feature.home.mvi.BalanceGameState
+import com.whatever.caramel.feature.home.mvi.BalanceGameCard
+import com.whatever.caramel.feature.home.mvi.BalanceGameOptionItem
 import com.whatever.caramel.feature.home.mvi.HomeState
-import kotlinx.collections.immutable.persistentListOf
+import kotlinx.collections.immutable.toImmutableList
 
 internal class QuizPreviewData : PreviewParameterProvider<HomeState> {
     override val values: Sequence<HomeState>
         get() {
             val myNickname = "내 닉네임"
-            val myGender = Gender.MALE
             val partnerNickname = "상대방 닉네임"
-            val partnerGender = Gender.FEMALE
-            val balanceGameState =
-                BalanceGameState(
-                    id = 0L,
-                    question = "밸런스 게임 퀴즈 질문",
-                    options =
-                        persistentListOf(
-                            BalanceGameOptionState(id = 0, name = "옵션 1"),
-                            BalanceGameOptionState(id = 1, name = "옵션 2"),
-                        ),
+            val balanceGameOptions =
+                listOf(
+                    BalanceGameOptionItem(id = 0, name = "옵션 1"),
+                    BalanceGameOptionItem(id = 1, name = "옵션 2"),
+                ).toImmutableList()
+            val balanceGameCard =
+                BalanceGameCard.initState().copy(
+                    question = "옵션1과 2중에 뭐가더 좋나요?",
+                    options = balanceGameOptions,
                 )
-            val myChoiceOption = balanceGameState.options[0]
-            val partnerChoiceOption = balanceGameState.options[1]
 
             return sequenceOf(
                 // 상대방 선택 x / 내 선택 x
                 HomeState(
                     myNickname = myNickname,
-                    balanceGameCardState = HomeState.BalanceGameCardState.IDLE,
-                    balanceGameState = balanceGameState,
-                    partnerChoiceOption = BalanceGameOptionState(),
-                    myChoiceOption = BalanceGameOptionState(),
+                    balanceGameCard =
+                        balanceGameCard.copy(
+                            myOption = null,
+                            partnerOption = null,
+                        ),
                 ),
                 // 상대방 선택 o / 내 선택 x
                 HomeState(
                     myNickname = myNickname,
                     partnerNickname = partnerNickname,
-                    balanceGameCardState = HomeState.BalanceGameCardState.IDLE,
-                    balanceGameState = balanceGameState,
-                    partnerChoiceOption = partnerChoiceOption,
-                    myChoiceOption = BalanceGameOptionState(),
+                    balanceGameCard =
+                        balanceGameCard.copy(
+                            myOption = null,
+                            partnerOption = balanceGameOptions[0],
+                        ),
                 ),
                 // 상대방 선택 x / 내 선택 o
                 HomeState(
                     myNickname = myNickname,
                     partnerNickname = partnerNickname,
-                    balanceGameCardState = HomeState.BalanceGameCardState.IDLE,
-                    balanceGameState = balanceGameState,
-                    partnerChoiceOption = BalanceGameOptionState(),
-                    myChoiceOption = myChoiceOption,
+                    balanceGameCard =
+                        balanceGameCard.copy(
+                            myOption = balanceGameOptions[0],
+                            partnerOption = null,
+                        ),
                 ),
                 // 상대방 선택 o / 내 선택 o
                 HomeState(
                     myNickname = myNickname,
                     partnerNickname = partnerNickname,
-                    balanceGameCardState = HomeState.BalanceGameCardState.IDLE,
-                    balanceGameState = balanceGameState,
-                    partnerChoiceOption = partnerChoiceOption,
-                    myChoiceOption = myChoiceOption,
+                    balanceGameCard =
+                        balanceGameCard.copy(
+                            myOption = balanceGameOptions[0],
+                            partnerOption = balanceGameOptions[1],
+                        ),
                 ),
                 // 결과 확인
                 HomeState(
                     myNickname = myNickname,
-                    myGender = myGender,
+                    myGender = Gender.MALE,
                     partnerNickname = partnerNickname,
-                    partnerGender = partnerGender,
-                    balanceGameCardState = HomeState.BalanceGameCardState.CONFIRM,
-                    balanceGameState = balanceGameState,
-                    partnerChoiceOption = partnerChoiceOption,
-                    myChoiceOption = myChoiceOption,
+                    partnerGender = Gender.FEMALE,
+                    isBalanceGameCardRotated = true,
+                    balanceGameCard =
+                        balanceGameCard.copy(
+                            myOption = balanceGameOptions[0],
+                            partnerOption = balanceGameOptions[0],
+                        ),
                 ),
             )
         }
